remove service type enum
